pyproject.toml
[project]
name = "adk-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[extensions]>=1.31.1",
    "httpx>=0.28",
    "litellm>=1.50",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.21",
    "asyncpg>=0.30",
    "authlib>=1.5",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["root_agent", "server"]

[project.scripts]
dev = "root_agent._cli:dev"
web = "root_agent._cli:web"
lint = "root_agent._cli:lint"
fmt = "root_agent._cli:fmt"
fix = "root_agent._cli:fix"
check = "root_agent._cli:check"
server = "root_agent._cli:server"

# ---------------------------------------------------------------------------
# Ruff 代码检查
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "F",      # pyflakes — 未使用导入、未定义变量等
    "E",      # pycodestyle 错误
    "W",      # pycodestyle 警告
    "I",      # isort — 导入排序
    "N",      # pep8-naming — 命名规范
    "UP",     # pyupgrade — 现代 Python 语法
    "B",      # flake8-bugbear — 常见陷阱
    "A",      # flake8-builtins — 内置名称遮蔽
    "T20",    # flake8-print — 禁止遗留 print()
    "SIM",    # flake8-simplify — 可简化表达式
    "TC",     # flake8-type-checking — TYPE_CHECKING 导入
    "RUF",    # ruff 专有规则
    "ANN",    # flake8-annotations — 强制类型注解
    "C4",     # flake8-comprehensions — 推导式优化
    "PIE",    # flake8-pie — 杂项检查
    "RET",    # flake8-return — 返回值一致性
    "PTH",    # flake8-use-pathlib — 优先 pathlib
]
ignore = [
    "ANN401",  # 禁止 typing.Any — 对 API 包装层过严
    "RUF001",  # 歧义 unicode — 中文文本误报
    "RUF002",  # 歧义 unicode（docstring 中）
    "RUF003",  # 歧义 unicode（注释中）
]

[tool.ruff.lint.isort]
known-first-party = ["root_agent", "gitea_agent", "model_config", "server"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"**/tools/*.py" = ["T20", "TC"]   # 允许 print()；ADK 需要运行时类型注解注入 ToolContext
"**/__init__.py" = ["F401"]       # 允许未使用导入（re-export）
"server/routers/*.py" = ["B008"]  # FastAPI Depends() 默认参数是惯用写法
"server/deps.py" = ["B008"]       # FastAPI Depends() 默认参数是惯用写法
"server/models/*.py" = ["TC003"]  # SQLAlchemy 映射列需要运行时 datetime
"server/schemas/*.py" = ["TC003"] # Pydantic 模型需要运行时 datetime
